+++ Data/Monoid/Cancellative.hs
@@ -1,5 +1,5 @@
 {- 
-    Copyright 2013-2017 Mario Blazevic
+    Copyright 2013-2019 Mario Blazevic
 
     License: BSD3 (see BSD3-LICENSE.txt file)
 -}
@@ -32,21 +32,21 @@
 -- 
 -- * 'RightGCDMonoid'
 
-{-# LANGUAGE Haskell2010, Trustworthy #-}
+{-# LANGUAGE Haskell2010, FlexibleInstances, Trustworthy #-}
 
 module Data.Monoid.Cancellative (
+   module Data.Semigroup.Cancellative,
    -- * Symmetric, commutative monoid classes
-   CommutativeMonoid, ReductiveMonoid, CancellativeMonoid, GCDMonoid(..),
+   CommutativeMonoid, ReductiveMonoid, CancellativeMonoid, MonoidWithMonus(..), GCDMonoid(..),
    -- * Asymmetric monoid classes
    LeftReductiveMonoid, RightReductiveMonoid,
    LeftCancellativeMonoid, RightCancellativeMonoid,
-   LeftGCDMonoid(..), RightGCDMonoid(..)
+   LeftGCDMonoid(..), RightGCDMonoid(..), OverlappingGCDMonoid(..)
    )
 where
 
 import qualified Prelude
 
-import Control.Applicative ((<$>), (<*>))
 import Data.Monoid -- (Monoid, Dual(..), Sum(..), Product(..))
 import qualified Data.ByteString as ByteString
 import qualified Data.ByteString.Unsafe as ByteString
@@ -60,6 +60,7 @@ import qualified Data.Sequence as Sequence
 import qualified Data.Set as Set
 import Data.Sequence (ViewL((:<)), ViewR((:>)), (<|), (|>))
 import qualified Data.Vector as Vector
+import Numeric.Natural (Natural)
 
 import Data.Semigroup.Cancellative
 
@@ -70,12 +71,18 @@ import Prelude hiding (gcd)
 -- > a <> b == b <> a
 class (Monoid m, CommutativeSemigroup m) => CommutativeMonoid m
 
--- | Class of Abelian monoids with a partial inverse for the Monoid '<>' operation. The inverse operation '</>' must
--- satisfy the following laws:
--- 
--- > maybe a (b <>) (a </> b) == a
--- > maybe a (<> b) (a </> b) == a
+-- | Class of Abelian monoids with a partial inverse for the '<>' operation.
 class (CommutativeMonoid m, ReductiveSemigroup m, LeftReductiveMonoid m, RightReductiveMonoid m) => ReductiveMonoid m
+
+-- | Class of Abelian monoids with monus. The monus operation '<\>' is a synonym for both 'stripPrefixOverlap' and
+-- 'stripSuffixOverlap', which must be equivalent as '<>' is both associative and commutative:
+--
+-- > (<\>) = flip stripPrefixOverlap
+-- > (<\>) = flip stripSuffixOverlap
+class (CommutativeMonoid m, OverlappingGCDMonoid m) => MonoidWithMonus m where
+   (<\>) :: m -> m -> m
+
+infix 5 <\>
 
 -- | Subclass of 'ReductiveMonoid' where '</>' is a complete inverse of the Monoid '<>' operation. The class instances
 -- must satisfy the following additional laws:
@@ -90,15 +97,15 @@ class (LeftCancellativeMonoid m, RightCancellativeMonoid m, ReductiveMonoid m) =
 -- > gcd a b == commonPrefix a b == commonSuffix a b
 -- > Just a' = a </> p && Just b' = b </> p
 -- >    where p = gcd a b
--- 
+--
 -- If a 'GCDMonoid' happens to also be a 'CancellativeMonoid', it should additionally satisfy the following laws:
--- 
+--
 -- > gcd (a <> b) (a <> c) == a <> gcd b c
 -- > gcd (a <> c) (b <> c) == gcd a b <> c
-class (ReductiveMonoid m, LeftGCDMonoid m, RightGCDMonoid m) => GCDMonoid m where
+class (ReductiveMonoid m, LeftGCDMonoid m, RightGCDMonoid m, OverlappingGCDMonoid m) => GCDMonoid m where
    gcd :: m -> m -> m
 
--- | Class of monoids with a left inverse of '<>', satisfying the following law:
+-- | Class of monoids with a left inverse of '<>', satisfying the following laws:
 -- 
 -- > isPrefixOf a b == isJust (stripPrefix a b)
 -- > maybe b (a <>) (stripPrefix a b) == b
@@ -108,7 +115,7 @@ class (ReductiveMonoid m, LeftGCDMonoid m, RightGCDMonoid m) => GCDMonoid m wher
 -- than linear in the length of the prefix argument.
 class (Monoid m, LeftReductiveSemigroup m) => LeftReductiveMonoid m
 
--- | Class of monoids with a right inverse of '<>', satisfying the following law:
+-- | Class of monoids with a right inverse of '<>', satisfying the following laws:
 -- 
 -- > isSuffixOf a b == isJust (stripSuffix a b)
 -- > maybe b (<> a) (stripSuffix a b) == b
@@ -124,7 +131,7 @@ class (Monoid m, RightReductiveSemigroup m) => RightReductiveMonoid m
 -- > stripPrefix a (a <> b) == Just b
 class (LeftCancellativeSemigroup m, LeftReductiveMonoid m) => LeftCancellativeMonoid m
 
--- | Subclass of 'LeftReductiveMonoid' where 'stripPrefix' is a complete inverse of '<>', satisfying the following
+-- | Subclass of 'RightReductiveMonoid' where 'stripSuffix' is a complete inverse of '<>', satisfying the following
 -- additional law:
 --
 -- > stripSuffix b (a <> b) == Just a
@@ -133,13 +140,22 @@ class (RightCancellativeSemigroup m, RightReductiveMonoid m) => RightCancellativ
 -- | Class of monoids capable of finding the equivalent of greatest common divisor on the left side of two monoidal
 -- values. The methods' complexity should be no worse than linear in the length of the common prefix. The following laws
 -- must be respected:
--- 
+--
 -- > stripCommonPrefix a b == (p, a', b')
 -- >    where p = commonPrefix a b
 -- >          Just a' = stripPrefix p a
 -- >          Just b' = stripPrefix p b
 -- > p == commonPrefix a b && p <> a' == a && p <> b' == b
 -- >    where (p, a', b') = stripCommonPrefix a b
+--
+-- Furthermore, 'commonPrefix' must return the unique greatest common prefix that contains, as its prefix, any other
+-- prefix @x@ of both values:
+--
+-- > not (x `isPrefixOf` a && x `isPrefixOf` b) || x `isPrefixOf` commonPrefix a b
+--
+-- and it cannot itself be a suffix of any other common prefix @y@ of both values:
+--
+-- > not (y `isPrefixOf` a && y `isPrefixOf` b && commonPrefix a b `isSuffixOf` y)
 class LeftReductiveMonoid m => LeftGCDMonoid m where
    commonPrefix :: m -> m -> m
    stripCommonPrefix :: m -> m -> (m, m, m)
@@ -162,6 +178,15 @@ class LeftReductiveMonoid m => LeftGCDMonoid m where
 -- >          Just b' = stripSuffix p b
 -- > s == commonSuffix a b && a' <> s == a && b' <> s == b
 -- >    where (a', b', s) = stripCommonSuffix a b
+--
+-- Furthermore, 'commonSuffix' must return the unique greatest common suffix that contains, as its suffix, any other
+-- suffix @x@ of both values:
+--
+-- > not (x `isSuffixOf` a && x `isSuffixOf` b) || x `isSuffixOf` commonSuffix a b
+--
+-- and it cannot itself be a prefix of any other common suffix @y@ of both values:
+--
+-- > not (y `isSuffixOf` a && y `isSuffixOf` b && commonSuffix a b `isPrefixOf` y)
 class RightReductiveMonoid m => RightGCDMonoid m where
    commonSuffix :: m -> m -> m
    stripCommonSuffix :: m -> m -> (m, m, m)
@@ -174,6 +199,32 @@ class RightReductiveMonoid m => RightGCDMonoid m where
             Just y' = stripSuffix s y
    {-# MINIMAL commonSuffix | stripCommonSuffix #-}
 
+-- | Class of monoids for which the greatest overlap can be found between any two values, such that
+--
+-- > a == a' <> overlap a b
+-- > b == overlap a b <> b'
+--
+-- The methods must satisfy the following laws:
+--
+-- > stripOverlap a b == (stripSuffixOverlap b a, overlap a b, stripPrefixOverlap a b)
+-- > stripSuffixOverlap b a <> overlap a b == a
+-- > overlap a b <> stripPrefixOverlap a b == b
+--
+-- The result of @overlap a b@ must be the largest prefix of @b@ and suffix of @a@, in the sense that it is contained
+-- in any other value @x@ that satifies the property @(x `isPrefixOf` b) && (x `isSuffixOf` a)@:
+--
+-- > (x `isPrefixOf` overlap a b) && (x `isSuffixOf` overlap a b)
+--
+-- and it must be unique so it's not contained in any other value @y@ that satisfies the same property @(y
+-- `isPrefixOf` b) && (y `isSuffixOf` a)@:
+--
+-- > not ((y `isPrefixOf` overlap a b) && (y `isSuffixOf` overlap a b) && y /= overlap a b)
+class (LeftReductiveMonoid m, RightReductiveMonoid m) => OverlappingGCDMonoid m where
+   stripPrefixOverlap :: m -> m -> m
+   stripSuffixOverlap :: m -> m -> m
+   overlap :: m -> m -> m
+   stripOverlap :: m -> m -> (m, m, m)
+
 -- Unit instances
 
 instance CommutativeMonoid ()
@@ -184,6 +235,9 @@ instance RightReductiveMonoid ()
 instance LeftCancellativeMonoid ()
 instance RightCancellativeMonoid ()
 
+instance MonoidWithMonus () where
+   () <\> () = ()
+
 instance GCDMonoid () where
    gcd () () = ()
 
@@ -192,6 +246,12 @@ instance LeftGCDMonoid () where
 
 instance RightGCDMonoid () where
    commonSuffix () () = ()
+
+instance OverlappingGCDMonoid () where
+   overlap () () = ()
+   stripOverlap () () = ((), (), ())
+   stripPrefixOverlap () () = ()
+   stripSuffixOverlap () () = ()
 
 -- Dual instances
 
@@ -206,11 +266,21 @@ instance RightCancellativeMonoid a => LeftCancellativeMonoid (Dual a)
 instance GCDMonoid a => GCDMonoid (Dual a) where
    gcd (Dual a) (Dual b) = Dual (gcd a b)
 
+instance MonoidWithMonus a => MonoidWithMonus (Dual a) where
+   Dual a <\> Dual b = Dual (a <\> b)
+
 instance LeftGCDMonoid a => RightGCDMonoid (Dual a) where
    commonSuffix (Dual a) (Dual b) = Dual (commonPrefix a b)
 
 instance RightGCDMonoid a => LeftGCDMonoid (Dual a) where
    commonPrefix (Dual a) (Dual b) = Dual (commonSuffix a b)
+
+instance OverlappingGCDMonoid a => OverlappingGCDMonoid (Dual a) where
+   overlap (Dual a) (Dual b) = Dual (overlap b a)
+   stripOverlap (Dual a) (Dual b) = (Dual s, Dual o, Dual p)
+      where (p, o, s) = stripOverlap b a
+   stripPrefixOverlap (Dual a) (Dual b) = Dual (stripSuffixOverlap a b)
+   stripSuffixOverlap (Dual a) (Dual b) = Dual (stripPrefixOverlap a b)
 
 -- Sum instances
 
@@ -222,14 +292,30 @@ instance Integral a => RightReductiveMonoid (Sum a)
 instance Integral a => LeftCancellativeMonoid (Sum a)
 instance Integral a => RightCancellativeMonoid (Sum a)
 
-instance (Integral a, Ord a) => GCDMonoid (Sum a) where
+instance {-# OVERLAPS #-} ReductiveMonoid (Sum Natural)
+instance {-# OVERLAPS #-} LeftReductiveMonoid (Sum Natural)
+instance {-# OVERLAPS #-} RightReductiveMonoid (Sum Natural)
+
+instance MonoidWithMonus (Sum Natural) where
+   Sum a <\> Sum b
+      | a > b = Sum (a - b)
+      | otherwise = Sum 0
+
+instance GCDMonoid (Sum Natural) where
    gcd (Sum a) (Sum b) = Sum (min a b)
 
-instance (Integral a, Ord a) => LeftGCDMonoid (Sum a) where
+instance LeftGCDMonoid (Sum Natural) where
    commonPrefix a b = gcd a b
 
-instance (Integral a, Ord a) => RightGCDMonoid (Sum a) where
+instance RightGCDMonoid (Sum Natural) where
    commonSuffix a b = gcd a b
+
+instance OverlappingGCDMonoid (Sum Natural) where
+   overlap a b = gcd a b
+   stripOverlap (Sum a) (Sum b) = (Sum $ a - c, Sum c, Sum $ b - c)
+      where c = min a b
+   stripPrefixOverlap = flip (<\>)
+   stripSuffixOverlap = flip (<\>)
 
 -- Product instances
 
@@ -238,14 +324,26 @@ instance Integral a => ReductiveMonoid (Product a)
 instance Integral a => LeftReductiveMonoid (Product a)
 instance Integral a => RightReductiveMonoid (Product a)
 
-instance Integral a => GCDMonoid (Product a) where
+instance MonoidWithMonus (Product Natural) where
+   Product 0 <\> Product 0 = Product 1
+   Product a <\> Product b = Product (a `div` Prelude.gcd a b)
+
+instance GCDMonoid (Product Natural) where
    gcd (Product a) (Product b) = Product (Prelude.gcd a b)
 
-instance Integral a => LeftGCDMonoid (Product a) where
+instance LeftGCDMonoid (Product Natural) where
    commonPrefix a b = gcd a b
 
-instance Integral a => RightGCDMonoid (Product a) where
+instance RightGCDMonoid (Product Natural) where
    commonSuffix a b = gcd a b
+
+instance OverlappingGCDMonoid (Product Natural) where
+   overlap a b = gcd a b
+   stripOverlap (Product 0) (Product 0) = (Product 1, Product 0, Product 1)
+   stripOverlap (Product a) (Product b) = (Product $ div a c, Product c, Product $ div b c)
+      where c = Prelude.gcd a b
+   stripPrefixOverlap = flip (<\>)
+   stripSuffixOverlap = flip (<\>)
 
 -- Pair instances
 
@@ -265,6 +363,14 @@ instance (LeftGCDMonoid a, LeftGCDMonoid b) => LeftGCDMonoid (a, b) where
 
 instance (RightGCDMonoid a, RightGCDMonoid b) => RightGCDMonoid (a, b) where
    commonSuffix (a, b) (c, d) = (commonSuffix a c, commonSuffix b d)
+
+instance (OverlappingGCDMonoid a, OverlappingGCDMonoid b) => OverlappingGCDMonoid (a, b) where
+   overlap (a1, b1) (a2, b2) = (overlap a1 a2, overlap b1 b2)
+   stripOverlap (a1, b1) (a2, b2) = ((ap, bp), (ao, bo), (as, bs))
+      where (ap, ao, as) = stripOverlap a1 a2
+            (bp, bo, bs) = stripOverlap b1 b2
+   stripPrefixOverlap (a1, b1) (a2, b2) = (stripPrefixOverlap a1 a2, stripPrefixOverlap b1 b2)
+   stripSuffixOverlap (a1, b1) (a2, b2) = (stripSuffixOverlap a1 a2, stripSuffixOverlap b1 b2)
 
 -- Triple instances
 
@@ -286,6 +392,16 @@ instance (LeftGCDMonoid a, LeftGCDMonoid b, LeftGCDMonoid c) => LeftGCDMonoid (a
 
 instance (RightGCDMonoid a, RightGCDMonoid b, RightGCDMonoid c) => RightGCDMonoid (a, b, c) where
    commonSuffix (a1, b1, c1) (a2, b2, c2) = (commonSuffix a1 a2, commonSuffix b1 b2, commonSuffix c1 c2)
+
+instance (OverlappingGCDMonoid a, OverlappingGCDMonoid b, OverlappingGCDMonoid c) =>
+         OverlappingGCDMonoid (a, b, c) where
+   overlap (a1, b1, c1) (a2, b2, c2) = (overlap a1 a2, overlap b1 b2, overlap c1 c2)
+   stripOverlap (a1, b1, c1) (a2, b2, c2) = ((ap, bp, cp), (ao, bo, co), (as, bs, cs))
+      where (ap, ao, as) = stripOverlap a1 a2
+            (bp, bo, bs) = stripOverlap b1 b2
+            (cp, co, cs) = stripOverlap c1 c2
+   stripPrefixOverlap (a1, b1, c1) (a2, b2, c2) = (stripPrefixOverlap a1 a2, stripPrefixOverlap b1 b2, stripPrefixOverlap c1 c2)
+   stripSuffixOverlap (a1, b1, c1) (a2, b2, c2) = (stripSuffixOverlap a1 a2, stripSuffixOverlap b1 b2, stripSuffixOverlap c1 c2)
 
 -- Quadruple instances
 
@@ -313,6 +429,19 @@ instance (LeftGCDMonoid a, LeftGCDMonoid b, LeftGCDMonoid c, LeftGCDMonoid d) =>
 instance (RightGCDMonoid a, RightGCDMonoid b, RightGCDMonoid c, RightGCDMonoid d) => RightGCDMonoid (a, b, c, d) where
    commonSuffix (a1, b1, c1, d1) (a2, b2, c2, d2) =
       (commonSuffix a1 a2, commonSuffix b1 b2, commonSuffix c1 c2, commonSuffix d1 d2)
+
+instance (OverlappingGCDMonoid a, OverlappingGCDMonoid b, OverlappingGCDMonoid c, OverlappingGCDMonoid d) =>
+         OverlappingGCDMonoid (a, b, c, d) where
+   overlap (a1, b1, c1, d1) (a2, b2, c2, d2) = (overlap a1 a2, overlap b1 b2, overlap c1 c2, overlap d1 d2)
+   stripOverlap (a1, b1, c1, d1) (a2, b2, c2, d2) = ((ap, bp, cp, dp), (ao, bo, co, dm), (as, bs, cs, ds))
+      where (ap, ao, as) = stripOverlap a1 a2
+            (bp, bo, bs) = stripOverlap b1 b2
+            (cp, co, cs) = stripOverlap c1 c2
+            (dp, dm, ds) = stripOverlap d1 d2
+   stripPrefixOverlap (a1, b1, c1, d1) (a2, b2, c2, d2) =
+      (stripPrefixOverlap a1 a2, stripPrefixOverlap b1 b2, stripPrefixOverlap c1 c2, stripPrefixOverlap d1 d2)
+   stripSuffixOverlap (a1, b1, c1, d1) (a2, b2, c2, d2) =
+      (stripSuffixOverlap a1 a2, stripSuffixOverlap b1 b2, stripSuffixOverlap c1 c2, stripSuffixOverlap d1 d2)
 
 -- Maybe instances
 
@@ -342,11 +471,20 @@ instance Ord a => LeftReductiveMonoid (Set.Set a)
 instance Ord a => RightReductiveMonoid (Set.Set a)
 instance Ord a => ReductiveMonoid (Set.Set a)
 
+instance Ord a => MonoidWithMonus (Set.Set a) where
+   (<\>) = (Set.\\)
+
 instance Ord a => LeftGCDMonoid (Set.Set a) where
    commonPrefix = Set.intersection
 
 instance Ord a => RightGCDMonoid (Set.Set a) where
    commonSuffix = Set.intersection
+
+instance Ord a => OverlappingGCDMonoid (Set.Set a) where
+   overlap = Set.intersection
+   stripOverlap a b = (Set.difference a b, Set.intersection a b, Set.difference b a)
+   stripPrefixOverlap a b = b <\> a
+   stripSuffixOverlap a b = b <\> a
 
 instance Ord a => GCDMonoid (Set.Set a) where
    gcd = Set.intersection
@@ -358,29 +496,52 @@ instance LeftReductiveMonoid IntSet.IntSet
 instance RightReductiveMonoid IntSet.IntSet
 instance ReductiveMonoid IntSet.IntSet
 
+instance MonoidWithMonus IntSet.IntSet where
+   (<\>) = (IntSet.\\)
+
 instance LeftGCDMonoid IntSet.IntSet where
    commonPrefix = IntSet.intersection
 
 instance RightGCDMonoid IntSet.IntSet where
    commonSuffix = IntSet.intersection
 
+instance OverlappingGCDMonoid IntSet.IntSet where
+   overlap = IntSet.intersection
+   stripOverlap a b = (IntSet.difference a b, IntSet.intersection a b, IntSet.difference b a)
+   stripPrefixOverlap a b = b <\> a
+   stripSuffixOverlap a b = b <\> a
+
 instance GCDMonoid IntSet.IntSet where
    gcd = IntSet.intersection
 
 -- Map instances
 
-instance Ord k => LeftReductiveMonoid (Map.Map k a)
+instance (Ord k, Eq a) => LeftReductiveMonoid (Map.Map k a)
+instance (Ord k, Eq a) => RightReductiveMonoid (Map.Map k a)
 
 instance (Ord k, Eq a) => LeftGCDMonoid (Map.Map k a) where
    commonPrefix = Map.mergeWithKey (\_ a b -> if a == b then Just a else Nothing) (const Map.empty) (const Map.empty)
 
+instance (Ord k, Eq v) => OverlappingGCDMonoid (Map.Map k v) where
+    overlap = Map.intersection
+    stripOverlap a b = (stripPrefixOverlap b a, overlap a b, stripSuffixOverlap a b)
+    stripPrefixOverlap = flip Map.difference
+    stripSuffixOverlap a b = Map.differenceWith (\x y-> if x == y then Nothing else Just x) b a
+
 -- IntMap instances
 
-instance LeftReductiveMonoid (IntMap.IntMap a)
+instance Eq a => LeftReductiveMonoid (IntMap.IntMap a)
+instance Eq a => RightReductiveMonoid (IntMap.IntMap a)
 
 instance Eq a => LeftGCDMonoid (IntMap.IntMap a) where
    commonPrefix = IntMap.mergeWithKey (\_ a b -> if a == b then Just a else Nothing)
-                                      (const IntMap.empty) (const IntMap.empty)
+                                       (const IntMap.empty) (const IntMap.empty)
+
+instance Eq a => OverlappingGCDMonoid (IntMap.IntMap a) where
+    overlap = IntMap.intersection
+    stripOverlap a b = (stripPrefixOverlap b a, overlap a b, stripSuffixOverlap a b)
+    stripPrefixOverlap = flip IntMap.difference
+    stripSuffixOverlap a b = IntMap.differenceWith (\x y-> if x == y then Nothing else Just x) b a
 
 -- List instances