+++ Data/Monoid/Instances/Stateful.hs
@@ -63,13 +63,9 @@ instance (Semigroup a, Semigroup b) => Semigroup (Stateful a b) where
    Stateful x <> Stateful y = Stateful (x <> y)
    {-# INLINE (<>) #-}
 
-instance (Semigroup a, Semigroup b) => Semigroup (Stateful a b) where
-   Stateful x <> Stateful y = Stateful (x <> y)
-   {-# INLINE (<>) #-}
-
 instance (Monoid a, Monoid b) => Monoid (Stateful a b) where
    mempty = Stateful mempty
-   mappend = (<>)
+   Stateful x `mappend` Stateful y = Stateful (mappend x y)
    {-# INLINE mempty #-}
    {-# INLINE mappend #-}